backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for the langchain-pgvector project"
authors = ["Thiara, Veer (V.) <vthiara+ford@ford.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.0"
sqlalchemy = "^2.0"
alembic = "^1.13"
psycopg2-binary = "^2.9"
pgvector = ">=0.3"
python-dotenv = "^1.0"
httpx = ">=0.27"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)"
]
